# packages/rest/src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"AnyJsonObject": {
				"description": "JSON object that can contain any key-value pairs",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
...
			"AnonCredsSchemaId": {
				"type": "string"
			},
			"AnonCredsSchema": {
				"properties": {
					"issuerId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"attrNames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"issuerId",
					"name",
					"version",
					"attrNames"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsGetSchemaSuccessResponse": {
				"properties": {
					"schemaId": {
						"$ref": "#/components/schemas/AnonCredsSchemaId"
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"resolutionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schemaId",
					"schema",
					"resolutionMetadata",
					"schemaMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Required_AnonCredsResolutionMetadata_": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"additionalProperties": {},
				"required": [
					"error",
					"message"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"AnonCredsGetSchemaFailedResponse": {
				"properties": {
					"schemaId": {
						"$ref": "#/components/schemas/AnonCredsSchemaId"
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"resolutionMetadata": {
						"$ref": "#/components/schemas/Required_AnonCredsResolutionMetadata_"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schemaId",
					"resolutionMetadata",
					"schemaMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateFinished": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"finished"
						],
						"nullable": false
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"schema",
					"schemaId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterSchemaSuccessResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"schemaState": {
						"$ref": "#/components/schemas/RegisterSchemaReturnStateFinished"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schemaState",
					"schemaMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterSchemaReturnStateFailed": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"failed"
						],
						"nullable": false
					},
					"reason": {
						"type": "string"
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterSchemaFailedResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"schemaState": {
						"$ref": "#/components/schemas/RegisterSchemaReturnStateFailed"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schemaState",
					"schemaMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegisterSchemaReturnStateWait": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"wait"
						],
						"nullable": false
					},
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"schemaId": {
						"type": "string"
					}
				},
				"required": [
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterSchemaWaitResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"schemaState": {
						"$ref": "#/components/schemas/RegisterSchemaReturnStateWait"
					},
					"schemaMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schemaState",
					"schemaMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterSchemaBody": {
				"properties": {
					"schema": {
						"$ref": "#/components/schemas/AnonCredsSchema"
					},
					"options": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"schema"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"schema": {
						"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
						"name": "schema-name",
						"version": "1.0",
						"attrNames": [
							"age"
						]
					}
				}
			},
			"AnonCredsCredentialDefinitionId": {
				"type": "string"
			},
			"AnonCredsCredentialDefinition": {
				"properties": {
					"issuerId": {
						"type": "string"
					},
					"schemaId": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"CL"
						],
						"nullable": false
					},
					"tag": {
						"type": "string"
					},
					"value": {
						"properties": {
							"revocation": {},
							"primary": {
								"$ref": "#/components/schemas/Record_string.unknown_"
							}
						},
						"required": [
							"primary"
						],
						"type": "object"
					}
				},
				"required": [
					"issuerId",
					"schemaId",
					"type",
					"tag",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsGetCredentialDefinitionSuccessResponse": {
				"properties": {
					"credentialDefinitionId": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinitionId"
					},
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinition"
					},
					"resolutionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"credentialDefinitionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"credentialDefinitionId",
					"credentialDefinition",
					"resolutionMetadata",
					"credentialDefinitionMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsGetCredentialDefinitionFailedResponse": {
				"properties": {
					"credentialDefinitionId": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinitionId"
					},
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinition"
					},
					"resolutionMetadata": {
						"$ref": "#/components/schemas/Required_AnonCredsResolutionMetadata_"
					},
					"credentialDefinitionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"credentialDefinitionId",
					"resolutionMetadata",
					"credentialDefinitionMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterCredentialDefinitionReturnStateFinished": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"finished"
						],
						"nullable": false
					},
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinition"
					},
					"credentialDefinitionId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"credentialDefinition",
					"credentialDefinitionId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionSuccessResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"credentialDefinitionState": {
						"$ref": "#/components/schemas/RegisterCredentialDefinitionReturnStateFinished"
					},
					"credentialDefinitionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"credentialDefinitionState",
					"credentialDefinitionMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterCredentialDefinitionReturnStateFailed": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"failed"
						],
						"nullable": false
					},
					"reason": {
						"type": "string"
					},
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinition"
					},
					"credentialDefinitionId": {
						"type": "string"
					}
				},
				"required": [
					"state",
					"reason"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionFailedResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"credentialDefinitionState": {
						"$ref": "#/components/schemas/RegisterCredentialDefinitionReturnStateFailed"
					},
					"credentialDefinitionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"credentialDefinitionState",
					"credentialDefinitionMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"RegisterCredentialDefinitionReturnStateWait": {
				"properties": {
					"state": {
						"type": "string",
						"enum": [
							"wait"
						],
						"nullable": false
					},
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsCredentialDefinition"
					},
					"credentialDefinitionId": {
						"type": "string"
					}
				},
				"required": [
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionWaitResponse": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"credentialDefinitionState": {
						"$ref": "#/components/schemas/RegisterCredentialDefinitionReturnStateWait"
					},
					"credentialDefinitionMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					},
					"registrationMetadata": {
						"$ref": "#/components/schemas/AnyJsonObject"
					}
				},
				"required": [
					"credentialDefinitionState",
					"credentialDefinitionMetadata",
					"registrationMetadata"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionInput": {
				"properties": {
					"issuerId": {
						"type": "string"
					},
					"schemaId": {
						"type": "string"
					},
					"tag": {
						"type": "string"
					}
				},
				"required": [
					"issuerId",
					"schemaId",
					"tag"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionOptions": {
				"properties": {
					"supportRevocation": {
						"type": "boolean"
					}
				},
				"required": [
					"supportRevocation"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRegisterCredentialDefinitionBody": {
				"properties": {
					"credentialDefinition": {
						"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionInput"
					},
					"options": {
						"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionOptions"
					}
				},
				"required": [
					"credentialDefinition",
					"options"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"credentialDefinition": {
						"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
						"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0",
						"tag": "definition"
					},
					"options": {
						"supportRevocation": true
					}
				}
			},
...
		},
		"securitySchemes": {
			"tenants": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "@credo-ts/rest",
		"version": "0.9.5",
		"description": "Rest API for using Credo over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/anoncreds/schemas/{schemaId}": {
			"get": {
				"operationId": "GetSchemaById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetSchemaSuccessResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {},
											"schemaMetadata": {},
											"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0",
											"schema": {
												"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
												"name": "schema-name",
												"version": "1.0",
												"attrNames": [
													"age"
												]
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid schemaId or unknown AnonCreds method provided",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetSchemaFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "Schema not found"
											},
											"schemaMetadata": {},
											"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Schema not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetSchemaFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "Schema not found"
											},
											"schemaMetadata": {},
											"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Unknown error retrieving schema",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetSchemaFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "Schema not found"
											},
											"schemaMetadata": {},
											"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve schema by schema id",
				"tags": [
					"AnonCreds"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "schemaId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AnonCredsSchemaId"
						}
					}
				]
			}
		},
		"/anoncreds/schemas": {
			"post": {
				"operationId": "RegisterSchema",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterSchemaSuccessResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"registrationMetadata": {},
											"schemaMetadata": {},
											"schemaState": {
												"state": "finished",
												"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0",
												"schema": {
													"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
													"name": "schema-name",
													"version": "1.0",
													"attrNames": [
														"string"
													]
												}
											}
										}
									}
								}
							}
						}
					},
					"202": {
						"description": "Wait for action to complete",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterSchemaWaitResponse"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Unknown error registering schema",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterSchemaFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"registrationMetadata": {},
											"schemaMetadata": {},
											"schemaState": {
												"state": "failed",
												"reason": "Unknown error occurred while registering schema",
												"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a new AnonCreds schema and registers the schema in the AnonCreds registry",
				"tags": [
					"AnonCreds"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AnonCredsRegisterSchemaBody"
							}
						}
					}
				}
			}
		},
		"/anoncreds/credential-definitions/{credentialDefinitionId}": {
			"get": {
				"operationId": "GetCredentialDefinitionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetCredentialDefinitionSuccessResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/CLAIM_DEF/20/definition",
											"credentialDefinition": {
												"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
												"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0",
												"type": "CL",
												"tag": "definition",
												"value": {
													"primary": {
														"n": "string",
														"s": "string",
														"r": {
															"master_secret": "string",
															"string": "string"
														},
														"rctxt": "string",
														"z": "string"
													},
													"revocation": {
														"g": "1 string",
														"g_dash": "string",
														"h": "string",
														"h0": "string",
														"h1": "string",
														"h2": "string",
														"htilde": "string",
														"h_cap": "string",
														"u": "string",
														"pk": "string",
														"y": "string"
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Invalid credentialDefinitionId or unknown AnonCreds method provided",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetCredentialDefinitionFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "CredentialDefinition not found"
											},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/CLAIM_DEF/20/definition"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "CredentialDefinition not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetCredentialDefinitionFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "CredentialDefinition not found"
											},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/CLAIM_DEF/20/definition"
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Unknown error retrieving credentialDefinition",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsGetCredentialDefinitionFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resolutionMetadata": {
												"error": "notFound",
												"message": "CredentialDefinition not found"
											},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/CLAIM_DEF/20/definition"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve credentialDefinition by credentialDefinition id",
				"tags": [
					"AnonCreds"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialDefinitionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AnonCredsCredentialDefinitionId"
						}
					}
				]
			}
		},
		"/anoncreds/credential-definitions": {
			"post": {
				"operationId": "RegisterCredentialDefinition",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionSuccessResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"registrationMetadata": {},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionState": {
												"state": "finished",
												"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/credentialDefinition-name/1.0",
												"credentialDefinition": {
													"issuerId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv",
													"schemaId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/schema-name/1.0",
													"type": "CL",
													"tag": "definition",
													"value": {
														"primary": {
															"n": "string",
															"s": "string",
															"r": {
																"master_secret": "string",
																"string": "string"
															},
															"rctxt": "string",
															"z": "string"
														},
														"revocation": {
															"g": "1 string",
															"g_dash": "string",
															"h": "string",
															"h0": "string",
															"h1": "string",
															"h2": "string",
															"htilde": "string",
															"h_cap": "string",
															"u": "string",
															"pk": "string",
															"y": "string"
														}
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"202": {
						"description": "Wait for action to complete",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionWaitResponse"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"500": {
						"description": "Unknown error registering credentialDefinition",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionFailedResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"registrationMetadata": {},
											"credentialDefinitionMetadata": {},
											"credentialDefinitionState": {
												"state": "failed",
												"reason": "Unknown error occurred while registering credentialDefinition",
												"credentialDefinitionId": "did:indy:bcovrin:test:WgWxqztrNooG92RXvxSTWv/anoncreds/v0/SCHEMA/credentialDefinition-name/1.0"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a new AnonCreds credentialDefinition and registers the credentialDefinition in the AnonCreds registry",
				"tags": [
					"AnonCreds"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AnonCredsRegisterCredentialDefinitionBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}